stargo: Starlark wrappers for all Go values and types
This is essentially Nate Finch's Starlight-Go idea,
but with an attempt to solve some of the aliasing problems
in the Starlark compiler by introducing the concept of
modes (addressable/var/lvalue versus nonaddressable/value/rvalue).

WORK IN PROGRESS -- contains bugs.

See stargo/cmd/stargo for something you can play with.